seperate model and train
