+++ toynlp/nnlm/README.md
@@ -1,0 +1,9 @@
+# NNLM
+
+## 数据集准备
+
+Salesforce/wikitext
+
+## Tokenizer训练
+
+## 模型训练
