+++ toynlp/nnlm/model.py
@@ -1,6 +1,4 @@
 import torch
-from datasets import load_dataset
-from torch.utils.data import DataLoader
 
 
 class NNLM(torch.nn.Module):
@@ -54,13 +52,3 @@ if __name__ == '__main__':
     )
     tokens = torch.randint(0, vocab_size, (2, 5))
     print(model(tokens).shape)
-    
-    # load dataset
-    def load_text_dataset():
-        """
-        https://huggingface.co/datasets/Salesforce/wikitext
-        https://huggingface.co/docs/transformers/perplexity#example-calculating-perplexity-with-gpt-2-in--transformers
-        """
-        ds = load_dataset("Salesforce/wikitext", "wikitext-2-v1")
-        dataloader = DataLoader(ds['train'], batch_size=32)
-        pass
